% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FoodGroup}
\alias{FoodGroup}
\title{Food Groups}
\format{
\subsection{\code{FoodGroup}}{

A data frame with XX rows and XX columns:
\describe{
\item{food_group_id}{XX}
\item{food_group_code}{XX}
\item{food_group_name}{XX}
\item{food_group_name_f}{XX}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
FoodGroup
}
\description{
XX This is more descriptive text about my data.
Even more info is provided here ...
}
\keyword{datasets}
