% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FoodNames}
\alias{FoodNames}
\title{Food Names}
\format{
\subsection{\code{FoodNames}}{

A data frame with 5691 rows and 10 columns:
\describe{
\item{food_id}{Primary key that is used to link the foods in other tables
to this one}
\item{food_group_id}{Foreign key from FOOD_GROUP to link with the food group's info}
\item{food_source_id}{Foreign key from FOOD_SOURCE to link with the source info}
\item{food_description}{Complete food name in English}
\item{food_description_f}{Complete food name in French}
\item{country_code}{Corresponds to the USDA NDB N/A code}
\item{food_date_of_entry}{Date the food name data was entered into the database}
\item{food_date_of_publication}{Date the food name data was originally published}
\item{scientific_name}{Scientific name of the food}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
FoodNames
}
\description{
This is a principal data frame. It stores information about each food in the
database. It contains a description of each food in English and French as
well as dates and comments.
}
\keyword{datasets}
