% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FoodSources}
\alias{FoodSources}
\title{Food Sources}
\format{
\subsection{\code{FoodSources}}{

A data frame with 17 rows and 4 columns:
\describe{
\item{food_source_id}{Primary key id used to link foods in other tables with
respective source}
\item{food_source_code}{Identifier code for the food sources. Food sources
give an indication as to Canadian content}
\item{food_source_description}{Food Source description in English}
\item{food_source_description_f}{Food Source description in French}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
FoodSources
}
\description{
This data frame is a support or "list" table that is used to link to the FOOD_NAME
table. It contains a list of several food sources (in English and French)
that foods can be grouped on.
}
\keyword{datasets}
