% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MeasureNames}
\alias{MeasureNames}
\title{Measure Names}
\format{
\subsection{\code{MeasureNames}}{

A data frame with 1163 rows and 3 columns:
\describe{
\item{measure_id}{Primary key id used to link foods to measures when using
conversion factor to measure nutrient, yield or refuse amounts}
\item{measure_name}{Measure description in English}
\item{measure_name_f}{Measure description in French}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
MeasureNames
}
\description{
This data frame is a support or "list" table that is used to link to the
CONVERSION_FACTOR table. It contains a list of measures (in English and French).
}
\keyword{datasets}
