% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NutrientAmounts}
\alias{NutrientAmounts}
\title{Nutrient Amounts}
\format{
\subsection{\code{NutrientAmounts}}{

A data frame with 524,675 rows and 7 columns:
\describe{
\item{food_id}{Foreign key from FOOD_NAME used to link to many tables in the database}
\item{nutrient_id}{Foreign key from NUTRIENT_NAME used to link to nutrient information }
\item{nutrient_value}{Mean value in 100g, edible portion. (The number of
decimal places does not reflect the accuracy of the data.).}
\item{standard_error}{Standard error of the mean}
\item{number_of_observations}{Number of samples}
\item{nutrient_source_id}{Identifier code generated by the database for
the nutrient source}
\item{nutrient_date_of_entry}{Date the data was entered into the database}

...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
NutrientAmounts
}
\description{
This is the main data frame. It uses information (by linking) from the FOOD_NAME
table (among others) to identify which nutrients and amounts are recorded
for that food.
}
\keyword{datasets}
