% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NutrientNames}
\alias{NutrientNames}
\title{Nutrient Names}
\format{
\subsection{\code{NutrientNames}}{

A data frame with 153 rows and 8 columns:
\describe{
\item{nutrient_id}{Primary key id used to link foods in other tables with
respective nutrients}
\item{nutrient_code}{Identifier code for nutrient name descriptions}
\item{nutrient_symbol}{Nutrient symbol or abbreviation of the nutrients.
They may differ from international nomenclature.}
\item{nutrient_unit}{Unit of measure (e.g., mg, g, mcg)}
\item{nutrient_name}{Nutrient name in English}
\item{nutrient_name_f}{Nutrient name in French}
\item{tagname}{International Network of Food Data Systems (INFOODS) Tagnames.
A unique Abbreviation for a food component developed by INFOODS to aid in
the interchange of data.
http://www.fao.org/infoods/infoods/standards-guidelines/food-component-identifiers-tagnames/en/}
\item{nutrient_decimals}{Number of decimal places used in the rounding of
the nutrient value}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
NutrientNames
}
\description{
This data frame is a support or "list" table that contains the list of nutrients
(in English and French) used in the NUTRIENT_AMOUNT table, with which it is linked.
}
\keyword{datasets}
