% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NutrientSources}
\alias{NutrientSources}
\title{Nutrient Sources}
\format{
\subsection{\code{NutrientSources}}{

A data frame with 21 rows and 4 columns:
\describe{
\item{nutrient_source_id}{Primary key id used to link nutrients in other
tables with respective source}
\item{nutrient_source_code}{Identifier code for nutrient source descriptions.}
\item{nutrient_source_description}{Nutrient Source description in English}
\item{nutrient_source_description_f}{Nutrient Source description in French}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
NutrientSources
}
\description{
This data frame is a support or "list" table that is used to link to the NUTRIENT_AMOUNT
table It contains a list of several sources and/or types of nutrient data
(in English and French).
}
\keyword{datasets}
