% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RefuseAmounts}
\alias{RefuseAmounts}
\title{Refuse Amounts}
\format{
\subsection{\code{RefuseAmounts}}{

A data frame with 6898 rows and 4 columns:
\describe{
\item{food_id}{The foreign key from FOOD_NAME that is used to determine what
food the particular refuse is for}
\item{refuse_id}{Foreign key from REFUSE_NAME used to connect foods to their
refuse amounts}
\item{refuse_amount}{The percent fo the portion that is refuse}
\item{refuse_date_of_entry}{Date the data was entered into the database}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
RefuseAmounts
}
\description{
This is a principal data frame. This table contains the percent of refuse, or
inedible portion, for each food.
}
\keyword{datasets}
