% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RefuseNames}
\alias{RefuseNames}
\title{Refuse Names}
\format{
\subsection{\code{RefuseNames}}{

A data frame with 151 rows and 3 columns:
\describe{
\item{refuse_id}{Primary key used to link foods to their corresponding
refuse type}
\item{refuse_name}{Refuse description in English}
\item{refuse_name_f}{Refuse description in French}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
RefuseNames
}
\description{
This file is a support or "list" table that is used to link to the REFUSE_AMOUNT
table. It contains a list of refuse types.
}
\keyword{datasets}
