% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YieldAmounts}
\alias{YieldAmounts}
\title{Yield Amounts}
\format{
\subsection{\code{YieldAmounts}}{

A data frame with 1427 rows and 4 columns:
\describe{
\item{food_id}{The foreign key from FOOD_NAME that is used to determine what
food the particular refuse is for}
\item{yield_id}{Foreign key from YIELD_NAME used to connect foods to their
yield amounts}
\item{yield_amount}{Yield from refuse and/or cooking losses}
\item{yield_date_of_entry}{Date the data was entered into the database}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
YieldAmounts
}
\description{
This is a principal data frame. This file table the yield from refuse and/or
cooking losses assigned to certain foods. These yields are most often used
for food inventory purposes.
}
\keyword{datasets}
