% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YieldNames}
\alias{YieldNames}
\title{Yield Names}
\format{
\subsection{\code{YieldNames}}{

A data frame with 80 rows and 3 columns:
\describe{
\item{yield_id}{Primary key used to link foods to their corresponding
yield type}
\item{yield_name}{Yield description in English}
\item{yield_name_f}{Yield description in French}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
YieldNames
}
\description{
This data frame is a support or "list" table that is used to link to the YIELD_AMOUNT
table. It contains a list of yield types or yield descriptions (in English and French).
}
\keyword{datasets}
