% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConversionFactor}
\alias{ConversionFactor}
\title{Conversion Factors}
\format{
\subsection{\code{ConversionFactor}}{

A data frame with XX rows and XX columns:
\describe{
\item{food_id}{XX}
\item{measure_id}{XX}
\item{conversion_factor_value}{XX}
\item{conv_factor_date_of_entry}{XX}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
ConversionFactor
}
\description{
XX This is more descriptive text about my data.
Even more info is provided here ...
}
\keyword{datasets}
