% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FoodGroup}
\alias{FoodGroup}
\title{Food Groups}
\format{
\subsection{\code{FoodGroup}}{

A data frame with 24 rows and 4 columns:
\describe{
\item{food_group_id}{The id (primary key) used to link food groups to the
Food_Name data frame}
\item{food_group_code}{Identifier code for the Canadian Nutrient File food
groups. There are 23 different CNF food groups}
\item{food_group_name}{Food group name in English}
\item{food_group_name_f}{Food group name in French}
...
}
}
}
\source{
\url{https://www.canada.ca/en/health-canada/services/food-nutrition/healthy-eating/nutrient-data/canadian-nutrient-file-2015-download-files.html}
}
\usage{
FoodGroup
}
\description{
This data frame is a support or "list" table that is used to link to the FOOD NAME
table. It contains a list of 23 different group headings (in English and
French) based on similar characteristics of the foods.
}
\keyword{datasets}
